Move extension methods for System.Object to their own namespace
